Remove unnecessary casts enabled by generics
Remove remaining unnecessary type casts in Savegame.java and ZonePoint.java
that were made obsolete by the addition of proper generic types to Hashtable
declarations.

Changes:
- Savegame.java: Removed (Integer) and (JavaFunction) casts
- ZonePoint.java: Removed (Double) casts

These casts are no longer needed because the generic types guarantee the
correct return types from the Hashtable.get() method.

+++ src/main/java/cgeo/geocaching/wherigo/openwig/ZonePoint.java
@@ -113,7 +113,7 @@ public class ZonePoint implements LuaTable, Serializable {
 
     public static double convertDistanceTo (double value, String unit) {
         if (unit != null && conversions.containsKey(unit)) {
-            return value / ((Double)conversions.get(unit)).doubleValue();
+            return value / conversions.get(unit).doubleValue();
         } else {
             return value;
         }
@@ -121,7 +121,7 @@ public class ZonePoint implements LuaTable, Serializable {
 
     public static double convertDistanceFrom (double value, String unit) {
         if (unit != null && conversions.containsKey(unit)) {
-            return value * ((Double)conversions.get(unit)).doubleValue();
+            return value * conversions.get(unit).doubleValue();
         } else {
             return value;
         }
